Added the ClassMap class and changed many other classes as a result.
The ClassMap class allows for binding of a single object to a single
Class instance (which holds the type of said object). This is a common
feature, previously implemented multiple times throughout the project
but is now nicely centralized. The ClassMap class is also capable of
remembering what order entries were added in, which is another feature
needed, and it provides access to a read-only list that contained the
objects in the order they are added if this option is enabled.