Made project almost completely independent from JavaFX.
- Only exception being the MasterRender which uses JavaFX's
GraphicsContext object to draw onto a Canvas.
- Removed initGui() method implementation from State and made it
abstract. Now, the State class represents the 'state' of a game in terms
of the GameSystems and the world being updated by the game. In addition,
because States are client-side only, a State also represents what is
being rendered from the game at any given time.
- Changed AClient to use java.util.Timer rather than the JavaFX Timeline
to further avoid using JavaFX-related API in this project.